chore: move internal/conf/validation to cmd/frontend
Only frontend uses this package, so moving this package here to make clearer what service depends on this code.

Test plan: Just moved a package, go compiler doesn't complain.
