+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -248,6 +248,7 @@ go_library(
         "//cmd/frontend/internal/auth/userpasswd",
         "//cmd/frontend/internal/cloneurls",
         "//cmd/frontend/internal/cody",
+        "//cmd/frontend/internal/conf/validation",
         "//cmd/frontend/internal/inventory",
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
@@ -269,7 +270,6 @@ go_library(
         "//internal/conf/conftypes",
         "//internal/conf/deploy",
         "//internal/conf/reposource",
-        "//internal/conf/validation",
         "//internal/database",
         "//internal/database/basestore",
         "//internal/database/migration",
