+++ cmd/frontend/internal/conf/validation/batches.go
