+++ cmd/frontend/internal/conf/validation/highlight.go
