+++ cmd/frontend/internal/conf/validation/licensing.go
