+++ cmd/frontend/internal/conf/validation/symbols.go
